+++ src/server/backend-protocol-handler.ts
@@ -94,6 +94,7 @@ export class BackendProtocolHandler {
 			case "files.upload.create": return this.api.createFileUpload(request.params);
 			case "files.upload.append": return this.api.appendFileUpload(request.params);
 			case "files.upload.complete": return this.api.completeFileUpload(request.params.uploadId);
+			case "files.upload.abort": return this.api.abortFileUpload(request.params.uploadId).then(() => ({}));
 			case "host.browse": return this.api.browseDirectory(request.params.path);
 			case "host.mkdir": return this.api.createDirectory(request.params.parentPath, request.params.name);
 			case "settings.get": return this.api.getLocalSettings();